Create node to store db version
