+++ api/src/graphql/connectors/index.js
@@ -1,60 +1,10 @@
-import _ from 'lodash';
 import uuidv4 from 'uuid/v4';
-
-// This fist connector works differently than the rest.
-// It does not get Nodes, but data records that can be from a calculation.
-// Because of this, it does not assign a __label propery.
-function runQueryAndGetRawData(session, query, params) {
-  return session.run(query, params)
-    .then((result) => {
-      session.close();
-      if (!result.records) return null;
-      return result.records[0].toObject();
-    });
-}
-
-function runQueryAndGetRecords(session, query, params) {
-  return session.run(query, params)
-    .then((result) => {
-      session.close();
-      if (!result.records) return null;
-      return result.records.map((r) => {
-        const { properties, labels } = r.get(0);
-        return Object.assign({}, properties, { __label: labels[0] });
-      });
-    });
-}
-
-function runQueryAndGetRecord(session, query, params) {
-  return runQueryAndGetRecords(session, query, params)
-    .then((records) => {
-      if (!records || records.length !== 1) return null;
-      return records[0];
-    });
-}
-
-function runQueryAndGetRecordsWithFields(session, query, params) {
-  return session.run(query, params)
-    .then((result) => {
-      session.close();
-      if (!result.records) return null;
-      return result.records.map((r) => {
-        const pairs = r.keys.map((key) => {
-          const { properties, labels } = r.get(key);
-          return [key, Object.assign({}, properties, { __label: labels[0] })];
-        });
-        return _.fromPairs(pairs);
-      });
-    });
-}
-
-function runQueryAndGetRecordWithFields(session, query, params) {
-  return runQueryAndGetRecordsWithFields(session, query, params)
-    .then((records) => {
-      if (!records || records.length !== 1) return null;
-      return records[0];
-    });
-}
+import {
+  runQueryAndGetRecords,
+  runQueryAndGetRecord,
+  runQueryAndGetRawData,
+  runQueryAndGetRecordWithFields,
+} from '../../db/cypherUtils';
 
 export function findNodesByLabel(driver, label) {
   const query = `
